+++ gs/frontend/mcc/src/utils/style-const.ts
